# docs/openapi.yaml
openapi: 3.0.3
info:
  title: OAuth 2.1 Authorization Server API
  description: |
    OAuth 2.1 (draft) compliant authorization server with PKCE, JWT access tokens,
    refresh token rotation, token introspection, and token revocation.
    
    Key features:
    - Authorization Code flow with mandatory PKCE (S256 only)
    - JWT access tokens signed with RS256
    - Refresh token rotation
    - Token introspection (RFC 7662)
    - Token revocation (RFC 7009)
    - OpenID Connect UserInfo endpoint
    - OpenID Connect Discovery
  version: 1.0.0
  contact:
    name: OAuth 2.1 Server
servers:
  - url: http://localhost:3000
    description: Local development server

tags:
  - name: Discovery
    description: OpenID Connect Discovery endpoints
  - name: Client Registration
    description: Dynamic client registration
  - name: Authorization
    description: Authorization code flow endpoints
  - name: Token
    description: Token endpoint (authorization_code and refresh_token grants)
  - name: Introspection
    description: Token introspection (RFC 7662)
  - name: Revocation
    description: Token revocation (RFC 7009)
  - name: UserInfo
    description: OpenID Connect UserInfo endpoint

paths:
  /.well-known/openid-configuration:
    get:
      tags: [Discovery]
      summary: OpenID Provider Configuration
      description: Returns the OpenID Provider configuration document.
      responses:
        '200':
          description: OpenID Provider configuration
          content:
            application/json:
              schema:
                type: object
                properties:
                  issuer:
                    type: string
                    example: http://localhost:3000
                  authorization_endpoint:
                    type: string
                    example: http://localhost:3000/authorize
                  token_endpoint:
                    type: string
                    example: http://localhost:3000/token
                  userinfo_endpoint:
                    type: string
                    example: http://localhost:3000/userinfo
                  jwks_uri:
                    type: string
                    example: http://localhost:3000/.well-known/jwks.json
                  introspection_endpoint:
                    type: string
                    example: http://localhost:3000/introspect
                  revocation_endpoint:
                    type: string
                    example: http://localhost:3000/revoke
                  registration_endpoint:
                    type: string
                    example: http://localhost:3000/register
                  response_types_supported:
                    type: array
                    items:
                      type: string
                    example: ["code"]
                  grant_types_supported:
                    type: array
                    items:
                      type: string
                    example: ["authorization_code", "refresh_token"]
                  code_challenge_methods_supported:
                    type: array
                    items:
                      type: string
                    example: ["S256"]
                  subject_types_supported:
                    type: array
                    items:
                      type: string
                    example: ["public"]
                  id_token_signing_alg_values_supported:
                    type: array
                    items:
                      type: string
                    example: ["RS256"]
                  token_endpoint_auth_methods_supported:
                    type: array
                    items:
                      type: string
                    example: ["client_secret_basic", "client_secret_post", "none"]
                  scopes_supported:
                    type: array
                    items:
                      type: string
                    example: ["openid", "profile", "email"]

  /.well-known/jwks.json:
    get:
      tags: [Discovery]
      summary: JSON Web Key Set
      description: Returns the JSON Web Key Set (JWKS) containing the public keys for verifying JWT signatures.
      responses:
        '200':
          description: JSON Web Key Set
          content:
            application/json:
              schema:
                type: object
                properties:
                  keys:
                    type: array
                    items:
                      type: object
                      properties:
                        kty:
                          type: string
                          example: RSA
                        n:
                          type: string
                          description: Modulus (base64url encoded)
                        e:
                          type: string
                          description: Exponent (base64url encoded)
                          example: AQAB
                        kid:
                          type: string
                          example: oauth21-key-1
                        alg:
                          type: string
                          example: RS256
                        use:
                          type: string
                          example: sig

  /register:
    post:
      tags: [Client Registration]
      summary: Register a new client
      description: |
        Register a new OAuth 2.1 client.
        Supports both public and confidential client types.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - client_name
                - client_type
                - redirect_uris
              properties:
                client_name:
                  type: string
                  description: Human-readable client name
                  example: My Test App
                client_type:
                  type: string
                  enum: [public, confidential]
                  description: |
                    Client type per OAuth 2.1.
                    Confidential clients receive a client_secret.
                  example: confidential
                redirect_uris:
                  type: array
                  items:
                    type: string
                    format: uri
                  description: |
                    Array of allowed redirect URIs.
                    Exact match is required (no wildcards).
                  example:
                    - https://example.com/callback
                grant_types:
                  type: array
                  items:
                    type: string
                    enum: [authorization_code, refresh_token]
                  default: [authorization_code, refresh_token]
      responses:
        '201':
          description: Client registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  client_id:
                    type: string
                    description: The client identifier
                  client_secret:
                    type: string
                    description: Client secret (only for confidential clients)
                    nullable: true
                  client_name:
                    type: string
                  client_type:
                    type: string
                    enum: [public, confidential]
                  redirect_uris:
                    type: array
                    items:
                      type: string
                  grant_types:
                    type: array
                    items:
                      type: string
                  created_at:
                    type: integer
                    description: Unix timestamp of registration
        '400':
          description: Invalid client metadata
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /authorize:
    get:
      tags: [Authorization]
      summary: Authorization endpoint
      description: |
        Starts the authorization code flow.
        PKCE (S256) is mandatory.
        Displays login and consent pages to the user.
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
          description: Must be "code" (only authorization code flow is supported)
          example: code
        - name: client_id
          in: query
          required: true
          schema:
            type: string
          description: The client identifier
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
            format: uri
          description: |
            The redirect URI. Must exactly match one of the registered redirect URIs.
        - name: scope
          in: query
          schema:
            type: string
          description: Space-separated list of scopes
          example: openid profile email
        - name: state
          in: query
          schema:
            type: string
          description: Opaque value used to maintain state between request and callback
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
          description: |
            PKCE code challenge (S256).
            Base64url-encoded SHA-256 hash of the code verifier.
        - name: code_challenge_method
          in: query
          schema:
            type: string
            enum: [S256]
            default: S256
          description: Code challenge method (only S256 is supported)
      responses:
        '302':
          description: Redirects to redirect_uri with authorization code or error
        '200':
          description: Login or consent page (HTML)
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /token:
    post:
      tags: [Token]
      summary: Token endpoint
      description: |
        Exchange authorization code for tokens, or refresh an access token.
        
        Supported grants:
        - authorization_code: Exchange authorization code for access and refresh tokens
        - refresh_token: Refresh an access token using a refresh token
        
        Refresh token rotation is enabled: every refresh_token grant issues a new refresh
        token and invalidates the old one.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              oneOf:
                - $ref: '#/components/schemas/AuthorizationCodeRequest'
                - $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Successful token response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          description: Invalid grant or request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid client authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /introspect:
    post:
      tags: [Introspection]
      summary: Token introspection (RFC 7662)
      description: |
        Determine whether a token is active (valid, not revoked, not expired).
        Client authentication is required.
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: The token to introspect
                token_type_hint:
                  type: string
                  enum: [access_token, refresh_token]
                  description: Hint about the token type
      security:
        - basicAuth: []
      responses:
        '200':
          description: Introspection response
          content:
            application/json:
              schema:
                type: object
                properties:
                  active:
                    type: boolean
                    description: Whether the token is active
                  scope:
                    type: string
                    description: Space-separated scopes (only if active)
                  client_id:
                    type: string
                    description: Client that the token was issued to
                  token_type:
                    type: string
                    description: Type of the token
                  exp:
                    type: integer
                    description: Expiration timestamp (Unix)
                  iat:
                    type: integer
                    description: Issued-at timestamp (Unix)
        '401':
          description: Invalid client authentication
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /revoke:
    post:
      tags: [Revocation]
      summary: Token revocation (RFC 7009)
      description: |
        Revoke a token (access or refresh).
        The endpoint always returns 200 OK even for invalid tokens
        (per RFC 7009).
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required:
                - token
              properties:
                token:
                  type: string
                  description: The token to revoke
                token_type_hint:
                  type: string
                  enum: [access_token, refresh_token]
                  description: Hint about the token type
      security:
        - basicAuth: []
      responses:
        '200':
          description: Token revoked (or token was already invalid)

  /userinfo:
    get:
      tags: [UserInfo]
      summary: Get user information
      description: |
        Returns information about the authenticated user.
        Requires a valid access token.
        Claims returned depend on the scopes granted.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                type: object
                properties:
                  sub:
                    type: string
                    description: Subject identifier (always present)
                  name:
                    type: string
                    description: Full name (profile scope)
                  preferred_username:
                    type: string
                    description: Username (profile scope)
                  email:
                    type: string
                    description: Email address (email scope)
                  email_verified:
                    type: boolean
                    description: Whether email is verified (email scope)
        '401':
          description: Invalid or missing access token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags: [UserInfo]
      summary: Get user information (POST)
      description: Same as GET /userinfo but accepts token via form body.
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                access_token:
                  type: string
      responses:
        '200':
          description: User information
          content:
            application/json:
              schema:
                type: object
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Client ID and client secret for confidential clients
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token for user endpoints

  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: OAuth 2.0 error code
          enum:
            - invalid_request
            - invalid_client
            - invalid_grant
            - invalid_scope
            - unauthorized_client
            - unsupported_grant_type
            - unsupported_response_type
            - access_denied
            - server_error
            - temporarily_unavailable
            - invalid_token
            - invalid_redirect_uri
            - invalid_client_metadata
        error_description:
          type: string
          description: Human-readable error description

    AuthorizationCodeRequest:
      type: object
      required:
        - grant_type
        - code
        - redirect_uri
        - code_verifier
      properties:
        grant_type:
          type: string
          enum: [authorization_code]
        code:
          type: string
          description: The authorization code received from /authorize
        redirect_uri:
          type: string
          format: uri
          description: Must match the redirect_uri used in the authorization request
        code_verifier:
          type: string
          description: PKCE code verifier (43-128 characters)
        client_id:
          type: string
          description: Required for public clients
        client_secret:
          type: string
          description: Required for confidential clients (alternative to Basic auth)

    RefreshTokenRequest:
      type: object
      required:
        - grant_type
        - refresh_token
      properties:
        grant_type:
          type: string
          enum: [refresh_token]
        refresh_token:
          type: string
          description: The refresh token
        scope:
          type: string
          description: |
            Optional subset of original scopes (space-separated).
            If omitted, uses original scopes.
        client_id:
          type: string
          description: Required for public clients
        client_secret:
          type: string
          description: Required for confidential clients (alternative to Basic auth)

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
          description: The access token (JWT, RS256 signed)
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Time until access token expires (seconds)
          example: 3600
        refresh_token:
          type: string
          description: The refresh token (for refresh_token grant)
        scope:
          type: string
          description: Space-separated list of granted scopes